+++ apps/sv/frontend/src/__tests__/mocks/constants.ts
@@ -110,6 +110,7 @@ export const svcInfo: GetSvcInfoResponse = {
             createFee: {
               fee: '0.03',
             },
+            extraFeaturedAppRewardAmount: '1.0',
             maxNumLockHolders: '50',
             transferFee: {
               initialRate: '0.01',